Add Day9
For the first part, I decided to do the things on the fly,
and deque really helped.

For the second part, I did not want to "hardcode" the solution
from the first part, so it is a bit messy. In any case, I ccould not
think of any optiomization (sorting or stripping the result...)
because it seems that the data can come in any form.