Day19
This is a small rework of https://github.com/AshKelly/advent-of-code-2020/blob/main/day19/solution.cpp
Check it out! The algorithm is not mine!
The blog is pretty cool! :) https://www.ashkelly.co.uk/blog/aoc20d19/

Differences:
    Maps are used because elements are not to be repeated.
    Deque insted of list because we don't want to do things in the middle of the container.
    Some C++17 stuff for if statements.
    idx used in function check to avoid copying strings around.
    Pattern preparation with stringstream and nested whiles.
    Some const references added to avoid copying things when not needed.

In any case, I was stuck with this one until I found AshKelly's solution because I spent a lot
of time trying to store a map of functions, and that killed me, so AshKelly's idea was very similar
to what I wanted (not a regex fan) but I miserably failed :( .